storage/fs: unexport SafeWriteToFile
Move, rename and unexport fs.SafeWriteToFile. Also, document its limitations.
This function is not broadly useful because the encryption-at-rest vfs.FS does
NOT provide atomic renames. This function can only be used safely with the
unencrypted filesystem.

Epic: none
Release note: none
