+++ pkg/storage/min_version.go
@@ -6,6 +6,7 @@
 package storage
 
 import (
+	"bytes"
 	"io"
 
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
@@ -42,7 +43,7 @@ func writeMinVersionFile(atomicRenameFS vfs.FS, dir string, version roachpb.Vers
 		return err
 	}
 	filename := atomicRenameFS.PathJoin(dir, MinVersionFilename)
-	if err := fs.SafeWriteToFile(atomicRenameFS, dir, filename, b, fs.UnspecifiedWriteCategory); err != nil {
+	if err := safeWriteToUnencryptedFile(atomicRenameFS, dir, filename, b, fs.UnspecifiedWriteCategory); err != nil {
 		return err
 	}
 	return nil
@@ -92,4 +93,45 @@ func getMinVersion(atomicRenameFS vfs.FS, dir string) (_ roachpb.Version, ok boo
 		return roachpb.Version{}, false, err
 	}
 	return version, true, nil
+}
+
+const tempFileExtension = ".crdbtmp"
+
+// safeWriteToUnencryptedFile writes the byte slice to the filename, contained
+// in dir, using the given fs. It returns after both the file and the containing
+// directory are synced.
+//
+// This function requires that the fs NOT be encrypted, because the
+// encryption-at-rest filesystem does NOT support atomic renames. See
+// pebble/vfs/atomicfs for a mechanism of atomically switching files on
+// encrypted filesystems.
+func safeWriteToUnencryptedFile(
+	fs vfs.FS, dir string, filename string, b []byte, category vfs.DiskWriteCategory,
+) error {
+	tempName := filename + tempFileExtension
+	f, err := fs.Create(tempName, category)
+	if err != nil {
+		return err
+	}
+	bReader := bytes.NewReader(b)
+	if _, err = io.Copy(f, bReader); err != nil {
+		f.Close()
+		return err
+	}
+	if err = f.Sync(); err != nil {
+		f.Close()
+		return err
+	}
+	if err = f.Close(); err != nil {
+		return err
+	}
+	if err = fs.Rename(tempName, filename); err != nil {
+		return err
+	}
+	fdir, err := fs.OpenDir(dir)
+	if err != nil {
+		return err
+	}
+	defer fdir.Close()
+	return fdir.Sync()
 }
